fixed load csv on initial page
refactor SiteCondition
