+++ src/features/stepSlice/api.js
@@ -30,8 +30,8 @@ export const getLocality = createAsyncThunk(
 
 export const getRegion = createAsyncThunk(
   "steps/getRegion",
-  async ({ regionId }, thunkAPI) => {
-    const url = `https://developapi.covercrop-selector.org/v2/regions/${regionId}`;
+  async ({ stateId }, thunkAPI) => {
+    const url = `https://developapi.covercrop-selector.org/v2/regions/${stateId}`;
     const res = await fetch(url).then((data) => data.json());
     return res;
   }
@@ -63,9 +63,7 @@ export const getSSURGOData = createAsyncThunk(
     const url = "https://sdmdataaccess.sc.egov.usda.gov/Tabular/post.rest";
     const res = await fetch(url, requestOptions).then((data) => data.json());
 
-    if (res !== {}) {
-      return res;
-    }
+    return res;
   }
 );
