+++ src/pages/Calculator/Steps/SiteCondition/LocationComponent/MapComponent/index.js
@@ -1,22 +1,62 @@
-import { Typography } from "@mui/material";
+import { useState, useEffect } from "react";
 import Grid from "@mui/material/Grid";
 import { Map } from "@psa/dst.ui.map";
 import { Button } from "@mui/material";
-import ArrowBackIosIcon from "@mui/icons-material/ArrowBackIos";
 import PlaceIcon from "@mui/icons-material/Place";
+import { useDispatch } from "react-redux";
+import {
+  getZoneData,
+  getSSURGOData,
+} from "../../../../../../features/stepSlice/api";
 
 import "./mapComponent.css";
-import { Dropdown } from "../../../../../../components/Dropdown";
 
 const MapComponent = ({
-  handleLocation,
-  lat,
-  lng,
   handleSteps,
   step,
-  selectedState,
-  states,
+  selectedToEditSite,
+  setSelectedToEditSite,
+  siteCondition,
+  handleUpdateSteps,
+  counties,
 }) => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const { latitude, longitude, address, zipCode, county } =
+      selectedToEditSite;
+
+    if (
+      latitude === siteCondition.latitude &&
+      longitude === siteCondition.longitude
+    ) {
+      return;
+    }
+
+    if (Object.keys(selectedToEditSite).length > 0) {
+      if (siteCondition.council === "MCCC") {
+        const filteredCounty = counties.filter((c) =>
+          county.toLowerCase().includes(c.label.toLowerCase())
+        );
+        if (filteredCounty.length > 0) {
+          console.log("filtered county", filteredCounty, county);
+          handleUpdateSteps("county", "siteCondition", filteredCounty[0].label);
+        }
+      }
+      handleUpdateSteps("latitude", "siteCondition", latitude);
+      handleUpdateSteps("longitude", "siteCondition", longitude);
+      handleUpdateSteps("address", "siteCondition", address);
+      handleUpdateSteps("zipCode", "siteCondition", zipCode);
+      dispatch(getZoneData({ zip: zipCode }));
+      dispatch(
+        getSSURGOData({
+          lat: latitude,
+          lon: longitude,
+        })
+      );
+    }
+  }, [selectedToEditSite]);
+
   return (
     <Grid xs={12} container className="map-container">
       <Grid xs={5} item>
@@ -24,7 +64,7 @@ const MapComponent = ({
           <Button
             className="location-back-button"
             variant="contained"
-            onClick={() => handleSteps("back", false)}
+            onClick={() => handleSteps("back")}
           >
             <PlaceIcon /> Select State
           </Button>
@@ -33,12 +73,12 @@ const MapComponent = ({
 
       <Grid xs={12} md={12} item>
         <Map
-          setAddress={handleLocation}
+          setAddress={setSelectedToEditSite}
           initWidth="100%"
           padding="20px"
           initHeight="360px"
-          initLat={lat}
-          initLon={lng}
+          initLat={siteCondition.latitude}
+          initLon={siteCondition.longitude}
           initStartZoom={12}
           initMinZoom={4}
           initMaxZoom={18}
